Scrape text from abstracts
An abstract does not contain text if the
title is 'ABSTRACT WITHDRAWN' or for an event such
as a reception or business meeting. For the text of these
abstracts, use an empty string.

An '\n' appears at the beginning and end of all
abstracts, and throughout in some abstracts. Save the raw
text and deal with these newline characters later.