Scrape date a second way
Scrape the date from the session web page, and ensure it is
the same as the date from the abstract web page. In the list
comprehension, "if date.text" accounts for abstracts like
EJ.00005 from 2019, and "date.text != 'Not Participating'"
accounts for abstracts like ED.00001 from 2021. The except clause
is for the Conference Experience for Undergraduates Poster Session.
This scrapes the date for years 2005-2021.