Save abstracts
Change to 'text = None' to differentiate between an empty
string that might actually be scraped and a placeholder for
when that information does not exist.

Write repr of abstract's text to file so the file always has six
lines (as mentioned earlier, an '\n' appears at the beginning
and end of all abstracts, and throughout in some abstracts).

Saving works for 2019, but will fail for other years.

Use 'if not abstract.{authors,text}' because this will account
for the empty string and None. An empty string will be scraped
for the author in 2019, for example EE.00003, EG.00003, and
EJ.00002. In other years (for example, 2021 ED.00001), scrape_abstract
will fail (for the date or authors) and raise an exception. If this
happens, as much information should be saved with the information
from scrape_session. In this case, authors would be None instead
of the empty string. So both (empty string and None) are accounted for.
Saving information from scrape_session will occur in the next commit.