Use session information when abstract information is unavailable
The function scrape_abstract will fail for certain abstracts from
years 2017, 2018, 2020, and 2021. Examples include 2017: CC.00003, and
2020: EM.00006. If the function fails, use the information from
the function scrape_session.

There appears to be a correlation between scrape_abstract failing
and the title 'Abstract Withdrawn' but not 'ABSTRACT WITHDRAWN'.
The lowercase version occurs for years 2017, 2018, 2020, and 2021, while
the uppercase version occurs for years in the interval [2006, 2020].